requirements.txt
requests
python-socketio
python-socketio[client]
paramiko